% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ACC.FUN}
\alias{ACC.FUN}
\title{Calculate the accuracy measures}
\usage{
ACC.FUN(yk, ck, wgtk, type = "FPR", u0 = 0.05)
}
\arguments{
\item{yk}{a vector of binary outcomes.}

\item{ck}{a vector of risk scores.}

\item{wgtk}{a vector of weights.}

\item{type}{a character which could be "FPR" or "TPR", and the default value is "FPR".}

\item{u0}{a numerical value for the accuracy measure specified in \code{type}, and the default value is 0.05.}
}
\value{
a vector of IPW estimates of the accuracy measures.
}
\description{
Calculate the (i) AUC, and (ii) TPR, PPV, and NPV at the cut-off value that leads to FPR being the given value, or FPR, PPV, and NPNV at the cut-off value that leads to TPR being the given value.
}
