% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{htWGT}
\alias{htWGT}
\title{Calculate the new weight by reformularizing the NCC sampling as a two-stage stratified sampling}
\usage{
htWGT(
  xi,
  di,
  id,
  case,
  control,
  m0,
  M.dat = NULL,
  a.M = NULL,
  yes.ptb = FALSE,
  control.list = list(CaseID = NULL, Vii.ptb = NULL, yes.parallel = FALSE)
)
}
\arguments{
\item{xi}{a vector of censored event times.}

\item{di}{a vector of binary values (1 or 0) indicating whether the true event time is observed.}

\item{id}{a vector which identifies the subjects.}

\item{case}{a vector which indicates whether each subject is a sub-cohort case. The value is 1 if the subject is a sub-cohort case, and it is 0 if the subject is not a sub-cohort case.}

\item{control}{a vector which indicates whether each subject is a sub-cohort control. The value is 1 if the subject is a sub-cohort control, and it is 0 if the subject is not a sub-cohort control.}

\item{m0}{a numerical value which is the number of controls selected for each case.}

\item{M.dat}{a \code{data.frame} which includes the matching variables. The default value is \code{NULL} if the selection is without matching.}

\item{a.M}{a numerical vector which includes the matching constraint for each matching variable. The default value is \code{NULL} if the selection is without matching.}

\item{yes.ptb}{logical indicating if obtaining the perturbed counterparts of the IPW estimates, and the default value is \code{FALSE}.}

\item{control.list}{a list including two elements: \code{CaseID} and \code{Vii.ptb} that control the perturbation procedure. \code{CaseID} is a vector in which listed the case that each subject that is selected in the sub-cohort for. If \code{id} is the same as \code{CaseID}, it indicates this subject is one of the sub-cohort cases. \code{Vii.ptb} is a vector of perturbation variables, and the default value is \code{NULL} if the perturbation procedure is not implemented for estimating the variance of the IPW estimators.}
}
\value{
If \code{yes.ptb=FALSE}, it returns a single-column matrix which includes the NCC weights; otherwise, it returns a matrix with the first column being the NCC weights and the remaining columns being their perturbed counterparts.
}
\description{
This proposed weight is the Horvitz-Thompson type of weight.
}
