% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysimu.R
\name{mysimu}
\alias{mysimu}
\title{Function to simulation NCC data}
\usage{
mysimu(N, yes.match = FALSE, m0, pi1)
}
\arguments{
\item{N}{sample size.}

\item{yes.match}{logical indicating whether the controls are selected with or without matching. The default value is \code{FALSE}.}

\item{m0}{a numerical value which is the number of controls selected for each case.}

\item{pi1}{the percentage of events that are selected as cases}
}
\value{
data a \code{data.frame} with the following variables: \code{time} (censored event time), \code{status} (censoring indicator), \code{marker1} and \code{marker2} (two markers), \code{id} (subjects' ID), \code{case} (indicator for sub-cohort case), \code{control} (indicator for sub-cohort control), \code{CaseID} (indicator which sub-cohort case that each subject is selected in the sub-cohort for).

If \code{yes.match=TRUE}, the list includes another two elements: \code{aM} (a vector of the matching constraint for each matching variable, i.e., \code{a0}) and \code{Mdat} (a \code{data.frame} including two matching variables).
}
\description{
Function to simulation NCC data
}
