% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{samWGT}
\alias{samWGT}
\title{Calculate the proposed weight which weights the event controls with the inverse selection probability}
\usage{
samWGT(xi, di, id, case, control, m0, M.dat = NULL, a.M = NULL)
}
\arguments{
\item{xi}{a vector of censored event times.}

\item{di}{a vector of binary values (1 or 0) indicating whether the true event time is observed.}

\item{id}{a vector which identifies the subjects.}

\item{case}{a vector which indicates whether each subject is a sub-cohort case. The value is 1 if the subject is a sub-cohort case, and it is 0 if the subject is not a sub-cohort case.}

\item{control}{a vector which indicates whether each subject is a sub-cohort control. The value is 1 if the subject is a sub-cohort control, and it is 0 if the subject is not a sub-cohort control.}

\item{m0}{a numerical value which is the number of controls selected for each case.}

\item{M.dat}{a \code{data.frame} which includes the matching variables. The default value is \code{NULL} if the selection is without matching.}

\item{a.M}{a numerical vector which includes the matching constraint for each matching variable. The default value is \code{NULL} if the selection is without matching.}
}
\value{
a vector of NCC weights.
}
\description{
This proposed weight modifies the method in Samuelsen (1997) by weighting the event controls with the inverse selection probability.
}
\references{
Samuelsen, S. (1997). A psudolikelihood approach to analysis of nested case-control studies. Biometrika, 84(2):379–394.
}
