% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{WGT.CEN}
\alias{WGT.CEN}
\title{Calculate the censoring weights}
\usage{
WGT.CEN(Ti, Di, t0, Vii.ptb = NULL)
}
\arguments{
\item{Ti}{a vector of censored event times.}

\item{Di}{a vector of binary values (1 or 0) indicating whether the true event time is observed.}

\item{t0}{a numerical value which is the prediction time period, used for creating the binary outcome status. The value is 1 if the event time is smaller than t0; otherwise, it is 0.}

\item{Vii.ptb}{a vector of perturbation variables. The default is \code{NULL} if the perturbation procedure is not implemented for estimating the variance of the IPW estimators.}
}
\value{
If \code{Vii.ptb=NULL}, it returns a single-column matrix which includes the censoring weights; otherwise, it returns a matrix with the first column being the censoring weights and the remaining columns being their perturbed counterparts.
}
\description{
Calculate the weights accounting for missing binary outcomes due to censoring.
}
