% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PH.IPW.R
\name{PH.IPW}
\alias{PH.IPW}
\title{IPW estimates for Cox PH model using the new weight}
\usage{
PH.IPW(
  formula,
  data,
  id,
  case,
  control,
  m0,
  t0,
  weight.type = "HT",
  yes.match = FALSE,
  control.matching = list(Mdat = NULL, aM = NULL),
  control.accuracy = list(type = "FPR", u0 = 0.05),
  yes.ptb = FALSE,
  control.ptb = list(seed = 123, n.ptb = 500, CaseID = NULL)
)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a \code{~} operator, and the terms on the right. The response must be a survival object as returned by the \code{Surv} function.}

\item{data}{a \code{data.frame} in which to interpret the variables named in the formula.}

\item{id}{character, which is the variable name in the \code{data} corresponding to the column that identifies the subjects.}

\item{case}{character, which is the variable name in the \code{data} corresponding to the column that indicates whether each subject is a sub-cohort case. The value is 1 if the subject is a sub-cohort case, and it is 0 if the subject is not a sub-cohort case.}

\item{control}{character, which is the variable name in the \code{data} corresponding to the column that indicates whether each subject is a sub-cohort control. The value is 1 if the subject is a sub-cohort control, and it is 0 if the subject is not a sub-cohort control. #' @param m0 a numerical value which is the number of controls selected for each case.}

\item{t0}{a numerical value, which is the prediction time period, used for creating the binary outcome status: =1 if the event time is smaller than \code{t0}, = 0 otherwise.}

\item{weight.type}{character which indicate which sampling weight to be used and whose value could be
\describe{
\item{"sam"}{the modified Samuelsen's weight by assigning the inverse selection probability to event cases;}
\item{"new"}{the modified Samuelsen's weight by assigning 0 to event cases;}
\item{"HT"}{the Horvitz-Thompson type of weight, which is the default value.}
}}

\item{yes.match}{logical indicating whether the controls are selected with or without matching; the default value is \code{FALSE}.}

\item{control.matching}{a list which includes two elements: \code{Mdat} and \code{aM}. \code{Mdat} is a \code{data.frame} which includes the matching variables, and its default value is \code{NULL} if the selection is without matching. \code{aM} is a numerical vector which includes the matching constraints for each matching variable, and its default value is \code{NULL} if the selection is without matching. The length of 'aM' should be the same as the number of columns of \code{Mdat}.}

\item{control.accuracy}{a list which includes the arguments used in estimating the accuracy measures. It includes two elements: \code{type} and \code{u0}. \code{type} is a character which could be "FPR" or "TPR", and the default value is "FPR". \code{u0} is a numerical value for the accuracy measure specified in \code{type}, and the default value is 0.05.}

\item{yes.ptb}{logical indicating if obtaining the perturbed counterparts of the IPW estimates, and the default value is \code{FALSE}.}

\item{control.ptb}{a list of elements that control the perturbation procedure. \code{seed} is a numerical value used in the function \code{\link{set.seed}}, and its default value is 123. \code{n.ptb} is a numerical value indicating the number of perturbed counterparts generated, and its default value is 100. \code{CaseID} is character, which is the variable name in the \code{data} corresponding to the column that listed the case that each subject that is selected in the sub-cohort for.}
}
\value{
a list of two elements: \code{estimates} and \code{ptb}. \code{estimates} is a list of IPW estimates of the model parameters (\code{coef}) and accuracy parameters (\code{acc}). If \code{yes.ptb=TRUE}, \code{ptb} is a list of the perturbed counterparts for the IPW estimates of the model parameters (\code{coef}) and accuracy parameters (\code{acc}); otherwise, \code{ptb = NULL}.
}
\description{
IPW estimates of model parameters and accuracy parameters for Cox's proportional hazards (PH) model with perturbed counterparts.
}
\examples{
\code{data("myexample")}
\code{PH.IPW(formula=Surv(time,status)~marker1+marker2,
      data=myexample$data,
      id="id",
      case="case",
      control="control",
      m0=3,t0=1,
      yes.match=T,control.matching=list(Mdat=myexample$Mdat,aM=myexample$aM),
      yes.ptb=TRUE,control.ptb=list(n.ptb=10,CaseID="CaseID"))}
}
